requirements.txt
streamlit
spotipy
pandas
python-dotenv
pydantic
tenacity
tzlocal
pytz
